Remove mqtt-broker-core
