+++ mqtt/Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "mqtt3",
     "mqtt-broker",
-    "mqtt-broker-core",
     "mqtt-edgehub",
     "mqtt-generic",
     "mqttd",
