+++ mqtt/mqtt-broker/Cargo.toml
@@ -14,6 +14,8 @@ fail = "0.3"
 flate2 = "1.0"
 futures = "0.3"
 futures-util = { version = "0.3", features = ["sink"] }
+humantime = "2.0"
+humantime-serde = "1.0"
 lazy_static = "1.4"
 openssl = "0.10"
 proptest = { version = "0.9", optional = true }
@@ -31,7 +33,6 @@ tracing-futures = "0.2"
 uuid = { version = "0.8", features = ["v4"] }
 
 mqtt3 = { path = "../mqtt3", features = ["serde1"] }
-mqtt-broker-core = { path = "../mqtt-broker-core" }
 
 [dev-dependencies]
 atty = "0.2"
