+++ mqtt/mqtt-broker/src/lib.rs
@@ -11,12 +11,14 @@
     clippy::missing_errors_doc
 )]
 
+mod auth;
 mod broker;
 mod connection;
 mod error;
 mod persist;
 mod server;
 mod session;
+mod settings;
 mod snapshot;
 mod state_change;
 mod subscription;
@@ -26,18 +28,21 @@ mod transport;
 #[cfg(any(test, feature = "proptest"))]
 pub mod proptest;
 
-use std::net::SocketAddr;
+use std::{
+    fmt::{Display, Formatter, Result as FmtResult},
+    net::SocketAddr,
+    sync::Arc,
+};
 
 use serde::{Deserialize, Serialize};
 use tokio::sync::OwnedSemaphorePermit;
 
 use mqtt3::proto;
-use mqtt_broker_core::{
-    auth::AuthId,
-    settings::{BrokerConfig, SessionConfig},
-    ClientId,
-};
 
+pub use crate::auth::{
+    authenticate_fn_ok, authorize_fn_ok, AuthId, AuthenticationContext, Authenticator,
+    Authorization, Authorizer, Certificate,
+};
 pub use crate::broker::{Broker, BrokerBuilder, BrokerHandle};
 pub use crate::connection::{
     ConnectionHandle, IncomingPacketProcessor, MakeIncomingPacketProcessor,
@@ -49,11 +54,56 @@ pub use crate::persist::{
 };
 pub use crate::server::Server;
 pub use crate::session::SessionState;
+pub use crate::settings::{BrokerConfig, SessionConfig};
 pub use crate::snapshot::{
     BrokerSnapshot, SessionSnapshot, ShutdownHandle, Snapshotter, StateSnapshotHandle,
 };
 pub use crate::subscription::{Segment, Subscription, TopicFilter};
 pub use crate::tls::ServerCertificate;
+
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+pub struct ClientId(Arc<String>);
+
+impl ClientId {
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl<T: Into<String>> From<T> for ClientId {
+    fn from(s: T) -> ClientId {
+        ClientId(Arc::new(s.into()))
+    }
+}
+
+impl Display for ClientId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        write!(f, "{}", self.as_str())
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ClientInfo {
+    peer_addr: SocketAddr,
+    auth_id: AuthId,
+}
+
+impl ClientInfo {
+    pub fn new(peer_addr: SocketAddr, auth_id: impl Into<AuthId>) -> Self {
+        Self {
+            peer_addr,
+            auth_id: auth_id.into(),
+        }
+    }
+
+    pub fn peer_addr(&self) -> SocketAddr {
+        self.peer_addr
+    }
+
+    pub fn auth_id(&self) -> &AuthId {
+        &self.auth_id
+    }
+}
 
 #[derive(Debug)]
 pub struct ConnReq {
