+++ mqtt/mqtt-broker/src/session/mod.rs
@@ -12,10 +12,11 @@ pub use state::SessionState;
 use std::collections::HashMap;
 
 use mqtt3::proto;
-use mqtt_broker_core::{auth::AuthId, ClientInfo};
 
-use crate::subscription::Subscription;
-use crate::{ClientEvent, ClientId, ConnReq, ConnectionHandle, Error};
+use crate::{
+    subscription::Subscription, AuthId, ClientEvent, ClientId, ClientInfo, ConnReq,
+    ConnectionHandle, Error,
+};
 
 #[derive(Debug)]
 pub enum Session {
