+++ mqtt/mqtt-edgehub/Cargo.toml
@@ -25,7 +25,6 @@ tracing = "0.1"
 url = "2.1"
 
 mqtt3 = { path = "../mqtt3", features = ["serde1"] }
-mqtt-broker-core = { path = "../mqtt-broker-core" }
 mqtt-broker = { path = "../mqtt-broker" }
 
 [target.'cfg(unix)'.dependencies]
